chore: Rename `serialize_with_tensorizer.py` to `tensorizer_utils.py`
