+++ exllamav2/config.py
@@ -187,7 +187,7 @@ class ExLlamaV2Config:
             with open(self.model_config, encoding = "utf8") as f:
                 read_config = json.load(f)
         else:
-            from util.serialize_with_tensorizer import read_stream
+            from util.tensorizer_utils import read_stream
             with read_stream(
                     os.path.join(self.model_dir, "config.json"),
                     **self.tensorizer_args) as stream:
@@ -344,7 +344,7 @@ class ExLlamaV2Config:
         # Even though this is lazy loaded, may be wasteful to have two TensorDeserializer instances
         if self.load_with_tensorizer:
             from tensorizer import TensorDeserializer
-            from util.serialize_with_tensorizer import read_stream
+            from util.tensorizer_utils import read_stream
 
             model_loc = self.model_dir or os.environ["TENSORIZER_LOC"]
             with read_stream(
