+++ exllamav2/model.py
@@ -284,7 +284,7 @@ class ExLlamaV2:
 
         if self.config.load_with_tensorizer:
             from tensorizer import TensorDeserializer
-            from util.serialize_with_tensorizer import read_stream
+            from util.tensorizer_utils import read_stream
             with read_stream(
                     os.path.join(self.config.model_dir, "model.tensors"),
                     **self.config.tensorizer_args) as stream:
